% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.gauss.quad-package.R
\docType{package}
\name{custom.gauss.quad-package}
\alias{custom.gauss.quad}
\alias{custom.gauss.quad-package}
\title{custom.gauss.quad: Custom Made Gauss Quadrature Nodes and Weights}
\description{
Use the high-precision arithmetic provided by the R package 'Rmpfr' to compute a custom-made Gauss quadrature nodes and weights, with up to 33 nodes, using a moment-based method via moment determinants. Paul Kabaila (2022) \href{https://arxiv.org/abs/2211.04729}{arXiv:2211.04729}.
}
\author{
\strong{Maintainer}: Paul Kabaila \email{p.kabaila@latrobe.edu.au} (\href{https://orcid.org/0000-0002-9205-668X}{ORCID})

}
\keyword{internal}
